backend/requirements.txt
# AI Photo Search System - Requirements
# Updated for Stage 1 completion with upgraded models + FastAPI Backend

# Core PyTorch ecosystem
torch>=2.0.0
torchvision>=0.15.0

# AI Models
transformers>=4.30.0              # LAION CLIP model
ultralytics>=8.0.0                # YOLOv8x object detection

# Computer Vision
opencv-python>=4.8.0              # Basic image processing
Pillow>=10.0.0                    # Image loading and EXIF

# Advanced Face Detection (Stage 1 upgrades)
insightface>=0.7.3                # State-of-the-art face detection
onnxruntime>=1.15.0               # Model inference engine
mtcnn>=0.1.1                      # Alternative face detection
facenet-pytorch>=2.5.3            # PyTorch face recognition

# Data Processing
numpy>=1.24.0,<2.0.0              # Numerical operations (NumPy 1.x for compatibility)
dateparser>=1.1.8                 # Temporal intelligence

# Utilities
matplotlib>=3.7.0                 # Optional visualization
requests>=2.28.0                  # Model downloads
scikit-learn>=1.3.0               # Clustering (DBSCAN) for Stage 2
networkx>=3.1                     # Relationship graph for Stage 2

# FastAPI Backend (NEW)
fastapi>=0.104.0                  # Modern web framework
uvicorn>=0.24.0                   # ASGI server
python-multipart>=0.0.6           # File upload support
pydantic>=2.4.0                   # Data validation
python-jose[cryptography]>=3.3.0  # JWT tokens (optional for auth)
httpx==0.28.1
pytest==8.4.2
# Version compatibility fixes
ml-dtypes==0.2.0                  # Fix for ONNX compatibility
onnx>=1.14.0,<1.15.0             # ONNX model format
tensorflow>=2.13.0,<2.16.0       # TensorFlow compatibility with NumPy 1.x
protobuf>=3.20.0,<5.0.0          # Protobuf compatibility
